Taking out StreamsInScopeIterator
